Update user serializer and view locations into accounts.
