+++ lfucg-exactions/server/accounts/views.py
@@ -1,13 +1,11 @@
-# from django.shortcuts import render
+from rest_framework.generics import RetrieveAPIView
 
-# from rest_framework.generics import RetrieveAPIView
+from django.contrib.auth.models import User
+from .serializers import UserSerializer
 
-# from django.contrib.auth.models import User
-# from .serializers import UserSerializer
+class CurrentUserDetails(RetrieveAPIView):
+    model = User
+    serializer_class = UserSerializer
 
-# class CurrentUserDetails(RetrieveAPIView):
-#     model = User
-#     serializer_class = UserSerializer
-
-#     def get_object(self):
-#         return self.request.user
+    def get_object(self):
+        return self.request.user
