+++ lfucg-exactions/server/api/views.py
@@ -1,11 +1,11 @@
-from rest_framework.generics import RetrieveAPIView
+# from rest_framework.generics import RetrieveAPIView
 
-from django.contrib.auth.models import User
-from .serializers import CurrentUserSerializer
+# from django.contrib.auth.models import User
+# from .serializers import CurrentUserSerializer
 
-class CurrentUserDetails(RetrieveAPIView):
-    model = User
-    serializer_class = CurrentUserSerializer
+# class CurrentUserDetails(RetrieveAPIView):
+#     model = User
+#     serializer_class = CurrentUserSerializer
 
-    def get_object(self):
-        return self.request.user
+#     def get_object(self):
+#         return self.request.user
